fn: Fix panic when a dynamic symbol is specified
If we can't find a symbol in the symbol table, we fall back to looking it up in
the dynamic symbol table. If it happens to be located there, we still try to
look up its name in the main symbol table, causing a panic when we call
unwrap().

Instead, look up the symbol in the same table from which it was obtained.